Frontend app

+++ app-backend/app/Domains/Profile/Repositories/ProfileRepository.php
@@ -4,6 +4,7 @@ namespace App\Domains\Profile\Repositories;
 use App\Domains\Profile\DataTransfers\ProfileDTO;
 use App\Domains\Profile\Models\Profile;
 use App\Domains\Profile\Repositories\Interfaces\ProfileRepositoryInterface;
+use App\Domains\Profile\Repositories\Interfaces\ProfileUserRepositoryInterface;
 use Illuminate\Http\Exceptions\HttpResponseException;
 use \Illuminate\Pagination\LengthAwarePaginator;
 
@@ -11,12 +12,15 @@ class ProfileRepository implements ProfileRepositoryInterface
 {
     private Profile $model;
 
+    private ProfileUserRepositoryInterface $profileUserRepository;
+
     /**
      * @param Profile $profile
      */
-    public function __construct(Profile $profile)
+    public function __construct(Profile $profile, ProfileUserRepositoryInterface $profileUserRepository)
     {
         $this->model = $profile;
+        $this->profileUserRepository = $profileUserRepository;
     }
 
     /**
@@ -71,6 +75,14 @@ class ProfileRepository implements ProfileRepositoryInterface
                 response()->json(['message' => 'perfil não encontrado'], 404)
             );
         }
+
+        $profileUserCounter = $this->profileUserRepository->countProfileUser($id);
+        if ($profileUserCounter > 0) {
+            throw new HttpResponseException(
+                response()->json(['message' => 'perfil não pode ser excluido, pois está em uso'], 403)
+            );
+        }
+
         return $model->delete();
     }
